requirements.txt
streamlit==1.29.0
pdf2docx==0.5.6
python-docx==1.0.1
PyPDF2==3.0.1
Pillow==10.1.0